feat: modify method location for reusability  (#66)
