+++ src/spaceone/cost_analysis/manager/job_manager.py
@@ -1,4 +1,5 @@
 import logging
+import copy
 from typing import List
 from datetime import datetime, timedelta
 
@@ -6,6 +7,8 @@ from spaceone.core.error import *
 from spaceone.core import queue, utils, config
 from spaceone.core.manager import BaseManager
 from spaceone.cost_analysis.model.job_model import Job
+from spaceone.cost_analysis.model.cost_model import CostQueryHistory
+from spaceone.cost_analysis.manager.cost_manager import CostManager
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -16,6 +19,7 @@ class JobManager(BaseManager):
         super().__init__(*args, **kwargs)
         self.job_model: Job = self.locator.get_model('Job')
         self.job_timeout = config.get_global('JOB_TIMEOUT', 7200)
+        self.cost_mgr: CostManager = self.locator.get_manager('CostManager')
 
     def is_job_running(self, data_source_id, domain_id):
         job_vos: List[Job] = self.job_model.filter(data_source_id=data_source_id, domain_id=domain_id,
@@ -81,6 +85,55 @@ class JobManager(BaseManager):
 
     def stat_jobs(self, query):
         return self.job_model.stat(**query)
+
+    def preload_cost_stat_queries(self, domain_id):
+        cost_query_cache_time = config.get_global('COST_QUERY_CACHE_TIME', 7)
+        cache_time = datetime.utcnow() - timedelta(days=cost_query_cache_time)
+
+        query = {
+            'filter': [
+                {'k': 'domain_id', 'v': domain_id, 'o': 'eq'},
+                {'k': 'updated_at', 'v': cache_time, 'o': 'gte'},
+            ]
+        }
+
+        _LOGGER.debug(f'[_preload_cost_stat_queries] cost_query_cache_time: {cost_query_cache_time} days')
+
+        history_vos, total_count = self.cost_mgr.list_cost_query_history(query)
+        for history_vo in history_vos:
+            _LOGGER.debug(f'[_preload_cost_stat_queries] create query cache: {history_vo.query_hash}')
+            self._create_cache_by_history(history_vo, domain_id)
+
+    def _create_cache_by_history(self, history_vo: CostQueryHistory, domain_id):
+        query = history_vo.query_options
+        query_hash = history_vo.query_hash
+        granularity = history_vo.granularity
+        start = history_vo.start
+        end = history_vo.end
+
+        # Original Date Range
+        self._create_cache(domain_id, copy.deepcopy(query), query_hash, granularity, start, end)
+
+    def _create_cache(self, domain_id, query, query_hash, granularity, start, end):
+        if granularity and start and end:
+            query = self.cost_mgr.add_date_range_filter(query, granularity, start, end)
+            query_hash_with_date_range = utils.dict_to_hash(query)
+
+            _LOGGER.debug(f'[_create_cache] query: {query}')
+
+            if 'group_by' in query:
+                if self.cost_mgr.is_monthly_cost(granularity, start, end):
+                    self.cost_mgr.analyze_monthly_costs_with_cache(query, query_hash, domain_id)
+                else:
+                    self.cost_mgr.analyze_costs_with_cache(query, query_hash, domain_id)
+            else:
+                if self.cost_mgr.is_monthly_cost(granularity, start, end):
+                    self.cost_mgr.stat_monthly_costs_with_cache(query, query_hash_with_date_range, domain_id,
+                                                                target='PRIMARY')
+                else:
+                    self.cost_mgr.stat_costs_with_cache(query, query_hash_with_date_range, domain_id, target='PRIMARY')
+        else:
+            self.cost_mgr.stat_monthly_costs_with_cache(query, query_hash, domain_id, target='PRIMARY')
 
     @staticmethod
     def decrease_remained_tasks(job_vo: Job):
